# doc-files/kypo-training-feedback-swagger-open-api.yaml
...
swagger: "2.0"
info:
  version: "Version: 1.0.0"
  title: "KYPO Training Feedback Service - API Reference"
host: "localhost:8082"
basePath: "/kypo-training-feedback/api/v1"
...
schemes:
  - "http"
  - "https"
paths:
...
  /trainees/sandboxes/{sandboxId}:
    get:
      tags:
        - "Trainees"
      summary: "Get trainee by sandboxId"
      description: ""
      operationId: "getTraineeInstance"
      produces:
        - "application/json"
      parameters:
        - name: "sandboxId"
          in: "path"
          description: "The trainee sandbox id"
          required: true
          type: "long"
          format: "int64"
      responses:
        200:
          description: "Trainee instance have been found."
          schema:
            $ref: "#/definitions/TraineeDTO"
        404:
          description: "Trainee instance have not been found."
          schema:
            $ref: "#/definitions/ApiError"
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
  /trainees/training-definitions/{definitionId}/training-instances/{instanceId}/sandboxes/{sandboxId}:
    post:
      tags:
        - "Trainees"
      summary: "Create trainee instance with computed cmdfeedback"
      description: "This should be done only once for every user"
      operationId: "createTraineeInstance"
      produces:
        - "application/json"
      parameters:
        - name: "definitionId"
          in: "path"
          description: "The training definition id"
          required: true
          type: "integer"
          format: "int64"
        - name: "instanceId"
          in: "path"
          description: "The training instance id"
          required: true
          type: "integer"
          format: "int64"
        - name: "sandboxId"
          in: "path"
          description: "The trainee sandbox id"
          required: true
          type: "integer"
          format: "int64"
      responses:
        200:
          description: "Trainee instance successfully created."
          schema:
            type: "object"
        404:
          description: "Training run with given id not found."
          schema:
            $ref: "#/definitions/ApiError"
        500:
          description: "Unexpected condition was encountered."
          schema:
            $ref: "#/definitions/ApiError"
definitions:
...
  ApiError:
    type: "object"
    properties:
      timestamp:
        type: "integer"
        format: "int64"
        example: "1574062900 (different for each type of exception)"
        description: "The time when the exception occurred"
      status:
        type: "string"
        example: "404 Not found (different for each type of exception)."
        description: "The HTTP response status code"
        enum:
          - "CONTINUE"
          - "SWITCHING_PROTOCOLS"
          - "PROCESSING"
          - "CHECKPOINT"
          - "OK"
          - "CREATED"
          - "ACCEPTED"
          - "NON_AUTHORITATIVE_INFORMATION"
          - "NO_CONTENT"
          - "RESET_CONTENT"
          - "PARTIAL_CONTENT"
          - "MULTI_STATUS"
          - "ALREADY_REPORTED"
          - "IM_USED"
          - "MULTIPLE_CHOICES"
          - "MOVED_PERMANENTLY"
          - "FOUND"
          - "MOVED_TEMPORARILY"
          - "SEE_OTHER"
          - "NOT_MODIFIED"
          - "USE_PROXY"
          - "TEMPORARY_REDIRECT"
          - "PERMANENT_REDIRECT"
          - "BAD_REQUEST"
          - "UNAUTHORIZED"
          - "PAYMENT_REQUIRED"
          - "FORBIDDEN"
          - "NOT_FOUND"
          - "METHOD_NOT_ALLOWED"
          - "NOT_ACCEPTABLE"
          - "PROXY_AUTHENTICATION_REQUIRED"
          - "REQUEST_TIMEOUT"
          - "CONFLICT"
          - "GONE"
          - "LENGTH_REQUIRED"
          - "PRECONDITION_FAILED"
          - "PAYLOAD_TOO_LARGE"
          - "REQUEST_ENTITY_TOO_LARGE"
          - "URI_TOO_LONG"
          - "REQUEST_URI_TOO_LONG"
          - "UNSUPPORTED_MEDIA_TYPE"
          - "REQUESTED_RANGE_NOT_SATISFIABLE"
          - "EXPECTATION_FAILED"
          - "I_AM_A_TEAPOT"
          - "INSUFFICIENT_SPACE_ON_RESOURCE"
          - "METHOD_FAILURE"
          - "DESTINATION_LOCKED"
          - "UNPROCESSABLE_ENTITY"
          - "LOCKED"
          - "FAILED_DEPENDENCY"
          - "UPGRADE_REQUIRED"
          - "PRECONDITION_REQUIRED"
          - "TOO_MANY_REQUESTS"
          - "REQUEST_HEADER_FIELDS_TOO_LARGE"
          - "UNAVAILABLE_FOR_LEGAL_REASONS"
          - "INTERNAL_SERVER_ERROR"
          - "NOT_IMPLEMENTED"
          - "BAD_GATEWAY"
          - "SERVICE_UNAVAILABLE"
          - "GATEWAY_TIMEOUT"
          - "HTTP_VERSION_NOT_SUPPORTED"
          - "VARIANT_ALSO_NEGOTIATES"
          - "INSUFFICIENT_STORAGE"
          - "LOOP_DETECTED"
          - "BANDWIDTH_LIMIT_EXCEEDED"
          - "NOT_EXTENDED"
          - "NETWORK_AUTHENTICATION_REQUIRED"
      message:
        type: "string"
        example: "The IDMGroup could not be found in database (different for each\
          \ type of exception)."
        description: "The specific description of the ApiError."
      errors:
        type: "array"
        example: "[The requested resource was not found (different for each type of\
          \ exception).]"
        description: "The list of main reasons of the ApiError."
        items:
          type: "string"
      path:
        type: "string"
        example: "/kypo2-rest-user-and-group/api/v1/groups/1000 (different for each\
          \ type of exception)."
        description: "The requested URI path which caused error."
    description: "Superclass for classes ApiEntityError and ApiMicroserviceError"
...
  TraineeDTO:
    type: "object"
    required:
      - "sandboxId"
    properties:
      sandboxId:
        type: "integer"
        format: "int64"
        example: 335
        description: "Id of trainee sandbox."
    description: "Trainee which absolved training."
...